requirements.txt
Django==1.11.7
djangorestframework==3.7.1
django-filter==1.1.0
django-extensions==1.9.7
ipython==6.2.1
psycopg2==2.7.3.2
Werkzeug==0.12.2
factory-boy==2.9.2
Faker==0.8.6
